Fixed getting the PORT of the current rocketchat installation in get_rocketchat_current_version() fixing issue-#37
In the function get_rocketchat_current_version() the PORT of the current rocketchat installation was not obtained correctly as the previous code depends on the location of the PORT definition in the corresponding line in /lib/systemd/system/rocketchat.service. As the location of the PORT in that string might vary the code must be able to obtain the PORT value regardless of the location. Fixing issue-#37